Time: 51 ms (15.67%), Space: 88.2 MB (74.56%) - LeetHub

+++ 1335-maximum-candies-allocated-to-k-children/1335-maximum-candies-allocated-to-k-children.cpp
@@ -20,7 +20,7 @@ public:
         for(auto x:candies)sum+=x;
 
         int start = 1;
-        int end = sum/k;
+        long long  end = sum/k;
 
 
         if(end==0) return 0;
